Split completion model into header and source
